test(benchmarks): bind planner corpus to package owners
Return the complete planner corpus to the approved CLI-integration path, bind its semantic globals and validation helper to tools.benchmark_plan after collection, move extraction-specific assertions into a focused package test, patch process classification against the compiler owner, and align the current Make help contract.
