+++ tests/boundary/process/tooling/test_benchmark_classification.py
@@ -8,12 +8,13 @@ from pathlib import Path
 from types import ModuleType, SimpleNamespace
 
 import pytest
+import tools.benchmark_plan.compiler as planner
 from tests.boundary.process.tooling.ci import ROOT
 
 _MISSING = object()
 
 
-def _load(module_state: pytest.MonkeyPatch) -> ModuleType:
+def _load_adapter(module_state: pytest.MonkeyPatch) -> ModuleType:
     path = ROOT / ".github/scripts/plan-benchmarks"
     loader = SourceFileLoader("plan_benchmarks", str(path))
     spec = importlib.util.spec_from_loader("plan_benchmarks", loader)
@@ -26,43 +27,47 @@ def _load(module_state: pytest.MonkeyPatch) -> ModuleType:
 
 
 @pytest.fixture
-def isolated_plan_benchmarks_module() -> Iterator[ModuleType]:
+def isolated_plan_benchmarks_adapter() -> Iterator[ModuleType]:
     previous = sys.modules.get("plan_benchmarks", _MISSING)
     with pytest.MonkeyPatch.context() as module_state:
-        yield _load(module_state)
+        yield _load_adapter(module_state)
     assert sys.modules.get("plan_benchmarks", _MISSING) is previous
 
 
-def test_loaded_planner_restores_module_state(
-    isolated_plan_benchmarks_module: ModuleType,
+def test_loaded_planner_adapter_restores_module_state(
+    isolated_plan_benchmarks_adapter: ModuleType,
 ) -> None:
-    assert sys.modules["plan_benchmarks"] is isolated_plan_benchmarks_module
+    assert sys.modules["plan_benchmarks"] is isolated_plan_benchmarks_adapter
+    assert isolated_plan_benchmarks_adapter.plan is planner.plan
 
 
-def _patch_plan(module: ModuleType, monkeypatch: pytest.MonkeyPatch) -> None:
+def _patch_plan(monkeypatch: pytest.MonkeyPatch) -> None:
     suite = SimpleNamespace(
         id="mathematical-benchmarks-v1",
         tasks=(SimpleNamespace(path=Path("autoformalization-semantic-audit")),),
     )
     monkeypatch.setattr(
-        module,
+        planner,
         "_membership",
         lambda: (
             {"autoformalization-semantic-audit": [(suite.id, suite.tasks[0].path)]},
             {suite.id: suite},
         ),
     )
-    monkeypatch.setattr(module, "_topology_digest", lambda suites: "sha256:" + "a" * 64)
-    monkeypatch.setattr(module, "_digest", lambda path: "sha256:" + "b" * 64)
+    monkeypatch.setattr(
+        planner,
+        "_topology_digest",
+        lambda suites: "sha256:" + "a" * 64,
+    )
+    monkeypatch.setattr(planner, "_digest", lambda path: "sha256:" + "b" * 64)
 
 
 def test_task_documentation_does_not_select_oracle(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    module = _load(monkeypatch)
-    _patch_plan(module, monkeypatch)
+    _patch_plan(monkeypatch)
 
-    plan = module.plan(
+    plan = planner.plan(
         [
             "benchmarks/datasets/mathematical-benchmarks-v1/"
             "autoformalization-semantic-audit/README.md"
@@ -84,10 +89,9 @@ def test_benchmark_control_tools_run_contract_gate_without_oracle(
     path: str,
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    module = _load(monkeypatch)
-    _patch_plan(module, monkeypatch)
+    _patch_plan(monkeypatch)
 
-    plan = module.plan([path], base="a" * 40, head="b" * 40)
+    plan = planner.plan([path], base="a" * 40, head="b" * 40)
 
     assert plan["run-benchmark-check"] == "true"
     assert plan["run-benchmark-record-schema"] == "true"
@@ -97,10 +101,9 @@ def test_benchmark_control_tools_run_contract_gate_without_oracle(
 def test_task_environment_selects_exact_task_oracle(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    module = _load(monkeypatch)
-    _patch_plan(module, monkeypatch)
+    _patch_plan(monkeypatch)
 
-    plan = module.plan(
+    plan = planner.plan(
         [
             "benchmarks/datasets/mathematical-benchmarks-v1/"
             "autoformalization-semantic-audit/environment/input.json"
@@ -117,16 +120,15 @@ def test_task_environment_selects_exact_task_oracle(
 def test_shared_environment_profile_escalates_only_on_integration_event(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    module = _load(monkeypatch)
-    _patch_plan(module, monkeypatch)
+    _patch_plan(monkeypatch)
 
-    pull_request = module.plan(
+    pull_request = planner.plan(
         ["benchmarks/environment-profiles.toml"],
         event="pull_request",
         base="a" * 40,
         head="b" * 40,
     )
-    merge_group = module.plan(
+    merge_group = planner.plan(
         ["benchmarks/environment-profiles.toml"],
         event="merge_group",
         base="a" * 40,
@@ -141,10 +143,9 @@ def test_shared_environment_profile_escalates_only_on_integration_event(
 def test_main_push_is_an_integration_owner(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    module = _load(monkeypatch)
-    _patch_plan(module, monkeypatch)
+    _patch_plan(monkeypatch)
 
-    push = module.plan(
+    push = planner.plan(
         ["benchmarks/environment-profiles.toml"],
         event="push",
         base="a" * 40,
